+++ test/fixtures/throwing-and.js
@@ -1,0 +1,4 @@
+Then(() => true);
+And(function () {
+  throw new Error('boom from the second And');
+});